+++ SmartWatchesStore/ViewModels/ProductFormViewModel.cs
@@ -4,14 +4,14 @@ namespace SmartWatchesStore.ViewModels
 {
     public class ProductFormViewModel
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         [Required, StringLength(250)]
         public string Name { get; set; }
         
         public string Color { get; set; }
 
-        public double Price { get; set; }
+        public decimal Price { get; set; }
 
         [Required, StringLength(2500)]
         public string Description { get; set; }
